refactor(adj): share the identical provenance-builder helpers
Each `build_adj_*_provenance.py` materializes one library's CAS bundle. They
legitimately differ in which bytes they cite and how those bytes decompose — but
three helpers had been copied verbatim between them.

A copied helper in a provenance builder is worse than ordinary duplication.
These functions decide the exact `start`/`end`/`quote_sha256` a claim is pinned
to, so two copies drifting apart would mean two libraries disagreeing about what
"cited" means while both reporting success. That is the failure this subsystem
exists to make impossible.

Extracted to `adj_provenance_builder.py`, chosen by AST comparison rather than by
eye:

  claim                 byte-identical across ALL FOUR builders
  input_claim_payload   byte-identical across the three dependent builders
  source_segments       byte-identical in percent-of and proportion; ratio
                        carried a variant WITHOUT the optional
                        `reasoned_discards` parameter. The larger version is a
                        strict superset — called with the default the loop never
                        runs — so ratio adopts it.

Deliberately NOT generalized: `build()`, `local_source()` and
`retained_external_source()` genuinely differ per library, and the four builders
split into two shapes that should not be forced together. A ROOT builder
(`arithmetic`) decomposes an external document into IR and is inherently
document-specific — §8's outstanding PDF decomposition for `percent.adj` is the
proof. DEPENDENT builders compose an already-verified root. Only provably
identical code moved.

VERIFIED BYTE-EXACTLY, which is the whole point. All four builders were re-run
against the real repository inside the Linux container and the checked-in CAS did
not move:

    arithmetic, ratio, percent-of, proportion  → rebuilt
    git status -- adj-stdlib-provenance        → 0 files changed
    verify.sh                                  → "valid": true

That oracle is stronger than a passing suite: not "nothing broke" but "the
content-addressed bytes are identical". It only became available with #9924 — CI
never re-runs these scripts, which is why the duplication went unnoticed and why
this refactor previously looked unverifiable.

Also repoints `test_adj_stdlib_provenance.py`, which reached into
`build_adj_ratio_provenance` by ATTRIBUTE for `input_claim_payload`. The call-site
rewrite covered the builder scripts and missed that caller, leaving
`test_transitive_imported_fact_witness_passes_strict_replay` — the guard that a
transitive imported-fact witness survives strict replay — dying on AttributeError
before its first assertion. Caught in review; the lesson is that a symbol's
consumers are not the files you happen to be editing.

Net -150 lines. Groundwork for the declarative migration driver: with the shared
floor in place, what remains per-library is genuinely per-library.

Tests (in-container, Linux): `test_adj_stdlib_provenance.py` 205 passed / 11
skipped; `test_build_adj_proportion_provenance.py` 6 passed; provenance verify
`valid: true`; ruff check and format clean on every touched file.
